Recipe book/requirements.txt
Django
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
dj-database-url
python-decouple
